+++ Kernel/System/Console/Command/Admin/Package/UpgradeAll.pm
@@ -57,15 +57,7 @@ sub Run {
         return $Self->ExitCodeOk();
     }
 
-    my $ConfigObject = $Kernel::OM->Get('Kernel::Config');
-    my %RepositoryList;
-    if ( $ConfigObject->Get('Package::RepositoryList') ) {
-        %RepositoryList = %{ $ConfigObject->Get('Package::RepositoryList') };
-    }
-
-    if ( $ConfigObject->Get('Package::RepositoryRoot') ) {
-        %RepositoryList = ( %RepositoryList, $PackageObject->PackageOnlineRepositories() );
-    }
+    my %RepositoryList = $PackageObject->_ConfiguredRepositoryDefinitionGet();
 
     # Show cloud repositories if system is registered.
     my $RepositoryCloudList;
@@ -73,7 +65,11 @@ sub Run {
         Key => 'Registration::State',
     ) || '';
 
-    if ( $RegistrationState eq 'registered' && !$ConfigObject->Get('CloudServices::Disabled') ) {
+    if (
+        $RegistrationState eq 'registered'
+        && !$Kernel::OM->Get('Kernel::Config')->Get('CloudServices::Disabled')
+        )
+    {
 
         $Self->Print("<yellow>Getting cloud repositories information...</yellow>\n");