+++ Kernel/System/Package.pm
@@ -5108,13 +5108,8 @@ sub _PackageOnlineListGet {
     my ( $Self, %Param ) = @_;
 
     my $ConfigObject = $Kernel::OM->Get('Kernel::Config');
-    my %RepositoryList;
-    if ( $ConfigObject->Get('Package::RepositoryList') ) {
-        %RepositoryList = %{ $ConfigObject->Get('Package::RepositoryList') };
-    }
-    if ( $ConfigObject->Get('Package::RepositoryRoot') ) {
-        %RepositoryList = ( %RepositoryList, $Self->PackageOnlineRepositories() );
-    }
+
+    my %RepositoryList = $Self->_ConfiguredRepositoryDefinitionGet();
 
     # Show cloud repositories if system is registered.
     my $RepositoryCloudList;
@@ -5124,7 +5119,6 @@ sub _PackageOnlineListGet {
 
     if ( $RegistrationState eq 'registered' && !$Self->{CloudServicesDisabled} ) {
         $RepositoryCloudList = $Self->RepositoryCloudList( NoCache => 1 );
-
     }
 
     my %RepositoryListAll = ( %RepositoryList, %{ $RepositoryCloudList || {} } );
@@ -5164,6 +5158,62 @@ sub _PackageOnlineListGet {
         PackageList   => \@PackageOnlineList,
         PackageLookup => \%PackageSoruceLookup,
     );
+}
+
+=head2 _ConfiguredRepositoryDefinitionGet()
+
+Helper function that gets the full list of configured package repositories updated for the current
+framework version.
+
+    my %RepositoryList = $PackageObject->_ConfiguredRepositoryDefinitionGet();
+
+Returns:
+
+    %RepositoryList = (
+        'http://ftp.otrs.org/pub/otrs/packages' => 'OTRS Free Features',
+        # ...,
+    );
+
+=cut
+
+sub _ConfiguredRepositoryDefinitionGet {
+    my ( $Self, %Param ) = @_;
+
+    my $ConfigObject = $Kernel::OM->Get('Kernel::Config');
+
+    my %RepositoryList;
+    if ( $ConfigObject->Get('Package::RepositoryList') ) {
+        %RepositoryList = %{ $ConfigObject->Get('Package::RepositoryList') };
+    }
+    if ( $ConfigObject->Get('Package::RepositoryRoot') ) {
+        %RepositoryList = ( %RepositoryList, $Self->PackageOnlineRepositories() );
+    }
+
+    return () if !%RepositoryList;
+
+    # Make sure ITSM repository matches the current framework version.
+    my @Matches = grep { $_ =~ m{http://ftp\.otrs\.org/pub/otrs/itsm/packages\d+/}msxi } sort keys %RepositoryList;
+
+    return %RepositoryList if !@Matches;
+
+    my @FrameworkVersionParts = split /\./, $Self->{ConfigObject}->Get('Version');
+    my $FrameworkVersion = $FrameworkVersionParts[0];
+
+    my $CurrentITSMRepository = "http://ftp.otrs.org/pub/otrs/itsm/packages$FrameworkVersion/";
+
+    # Delete all old ITSM repositories, but leave the current if exists
+    for my $Repository (@Matches) {
+        if ( $Repository ne $CurrentITSMRepository ) {
+            delete $RepositoryList{$Repository};
+        }
+    }
+
+    return %RepositoryList if exists $RepositoryList{$CurrentITSMRepository};
+
+    # Make sure that current ITSM repository is in the list.
+    $RepositoryList{$CurrentITSMRepository} = 'OTRS::ITSM 6 Master';
+
+    return %RepositoryList;
 }
 
 sub DESTROY {